+++ client/package.json
@@ -37,6 +37,16 @@
     "@radix-ui/react-toggle": "^1.1.2",
     "@radix-ui/react-toggle-group": "^1.1.2",
     "@radix-ui/react-tooltip": "^1.1.8",
+    "@tiptap/extension-image": "^2.10.2",
+    "@tiptap/extension-table": "^2.10.2",
+    "@tiptap/extension-table-cell": "^2.10.2",
+    "@tiptap/extension-table-header": "^2.10.2",
+    "@tiptap/extension-table-row": "^2.10.2",
+    "@tiptap/extension-task-item": "^2.10.2",
+    "@tiptap/extension-task-list": "^2.10.2",
+    "@tiptap/pm": "^2.10.2",
+    "@tiptap/react": "^2.10.2",
+    "@tiptap/starter-kit": "^2.10.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
@@ -46,7 +56,6 @@
     "lucide-react": "^0.487.0",
     "next": "15.0.3",
     "next-themes": "^0.4.6",
-    "quill": "^2.0.3",
     "react": "19.0.0-rc-66855b96-20241106",
     "react-day-picker": "^8.10.1",
     "react-dom": "19.0.0-rc-66855b96-20241106",
@@ -58,12 +67,12 @@
     "sonner": "^2.0.3",
     "tailwind-merge": "^3.1.0",
     "tailwindcss-animate": "^1.0.7",
+    "tiptap-extension-resize-image": "^1.2.1",
     "vaul": "^1.1.2",
     "zod": "^3.24.2"
   },
   "devDependencies": {
     "@types/node": "^20",
-    "@types/quill": "^2.0.14",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "eslint": "^8",
