+++ client/src/app/globals.css
@@ -18,55 +18,9 @@ body {
   margin: 0;
 }
 
-.container .ql-editor {
-  border: none;
-  width: 210mm !important;
-  min-height: 297mm !important;
-  padding: 0.5in !important;
-  margin: 1rem;
-  box-shadow: 0 0 5px 0 rgba(0, 0, 0, 0.5);
-  background-color: white;
-}
-
-.container .ql-container.ql-snow {
-  border: none;
-  display: flex;
-  justify-content: center;
-  background-color: hsl(var(--primary-foreground));
-  padding-top: 0.5rem;
-}
-
-.container .ql-toolbar.ql-snow {
-  display: flex;
-  justify-content: center;
-  position: sticky;
-  top: 3.5rem;
-  z-index: 1;
-  background-color: hsl(var(--background));
-  border: none;
-  box-shadow: 0 0 5px 0 rgba(0, 0, 0, 0.5);
-  width: 100vw;
-}
-
-@page {
-  margin: 0.5in;
-}
-
 @media print {
   body {
     background: none;
-  }
-
-  .container .ql-editor {
-    width: calc(210mm - 1in);
-    height: calc(297mm - 1in);
-    padding: 0;
-    margin: 0;
-    align-self: flex-start;
-    justify-self: center;
-  }
-  .container .ql-toolbar.ql-snow {
-    display: none;
   }
 
   div.os-scrollbar,
@@ -153,5 +107,172 @@ body {
   }
   body {
     @apply bg-background text-foreground;
+  }
+}
+
+.tiptap {
+  /* Image-specific styling */
+  img {
+    display: block;
+    height: auto;
+    margin: 1.5rem 0;
+    max-width: 100%;
+
+    &.ProseMirror-selectednode {
+      outline: 3px solid var(--primary);
+    }
+  }
+
+  /* Table-specific styling */
+  table {
+    border-collapse: collapse;
+    margin: 0;
+    overflow: hidden;
+    table-layout: fixed;
+    width: 100%;
+
+    td,
+    th {
+      border: 1px solid black;
+      box-sizing: border-box;
+      min-width: 1em;
+      padding: 6px 8px;
+      position: relative;
+      vertical-align: top;
+
+      > * {
+        margin-bottom: 0;
+      }
+    }
+
+    th {
+      background-color: #c7c7c7;
+      font-weight: bold;
+      text-align: left;
+    }
+
+    .selectedCell:after {
+      background: rgba(0, 0, 0, 0.2);
+      content: "";
+      left: 0;
+      right: 0;
+      top: 0;
+      bottom: 0;
+      pointer-events: none;
+      position: absolute;
+      z-index: 2;
+    }
+
+    .column-resize-handle {
+      background-color: var(--primary);
+      bottom: -2px;
+      pointer-events: none;
+      position: absolute;
+      right: -2px;
+      top: 0;
+      width: 4px;
+    }
+  }
+
+  .tableWrapper {
+    margin: 1.5rem 0;
+    overflow-x: auto;
+  }
+
+  &.resize-cursor {
+    cursor: ew-resize;
+    cursor: col-resize;
+  }
+
+  /* Heading styles */
+  h1,
+  h2,
+  h3,
+  h4,
+  h5,
+  h6 {
+    line-height: 1.1;
+    margin-top: 2.5rem;
+    text-wrap: pretty;
+  }
+
+  h1,
+  h2 {
+    margin-top: 3.5rem;
+    margin-bottom: 1.5rem;
+  }
+
+  h1 {
+    font-size: 1.4rem;
+  }
+
+  h2 {
+    font-size: 1.2rem;
+  }
+
+  h3 {
+    font-size: 1.1rem;
+  }
+
+  h4,
+  h5,
+  h6 {
+    font-size: 1rem;
+  }
+
+  /* List styles */
+  ul,
+  ol {
+    padding: 0 1rem;
+    margin: 1.25rem 1rem 1.25rem 0.4rem;
+  }
+
+  ul li {
+    list-style-type: disc;
+
+    p {
+      margin-top: 0.25em;
+      margin-bottom: 0.25em;
+    }
+  }
+
+  ol li {
+    list-style-type: decimal;
+    p {
+      margin-top: 0.25em;
+      margin-bottom: 0.25em;
+    }
+  }
+
+  /* Task list specific styles */
+  ul[data-type="taskList"] {
+    list-style: none;
+    margin-left: 0;
+    padding: 0;
+
+    li {
+      align-items: flex-start;
+      display: flex;
+
+      > label {
+        flex: 0 0 auto;
+        margin-right: 0.5rem;
+        user-select: none;
+      }
+
+      > div {
+        flex: 1 1 auto;
+      }
+    }
+
+    input[type="checkbox"] {
+      cursor: pointer;
+      margin-top: 0.6rem;
+      accent-color: black;
+    }
+
+    ul[data-type="taskList"] {
+      margin: 0;
+    }
   }
 }
